+++ nw/ARENA/OBASE/Vehicle/VehicleRuntimeState.cpp
@@ -733,6 +733,7 @@ bool VehicleRuntimeState_ApplyControlAt(
     case TURN_LEFT:
     case TURN_RIGHT:
     case FIRE_PRIMARY:
+    case JUMP:
     case STOP_VEHICLE:
     case CHANGE_VEHICLE:
         break;